requirements.txt
absl-py==0.9.0
asn1crypto==0.24.0
astor==0.8.1
atari-py==0.2.6
atomicwrites==1.2.1
attrs==19.3.0
beautifultable==1.0.0
boto3==1.9.57
botocore==1.12.253
box2d-py==2.3.8
cachetools==3.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle==0.6.1
colorama==0.4.3
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.17
-e git+https://github.com/rail-berkeley/d4rl.git@c4bd3dea1e35920db3dc72fe5ef3c02c9be10d4b#egg=d4rl
dask==1.0.0
decorator==4.4.2
deepdiff==3.3.0
# Editable install with no version control (dm-control==0.0.340626854)
# -e /NAS2020/Share/mcchen/mopo/dm_control
dm-control==0.0.340626854
dm-env==1.3
dm-tree==0.1.5
docutils==0.15.2
dotmap==1.3.8
filelock==3.0.12
Flask==1.0.2
flatbuffers==1.12
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gitdb2==2.0.5
GitPython==2.1.11
glfw==1.11.0
gpflow==1.4.1
grpcio==1.28.1
gtimer==1.0.0b5
gym==0.12.0
h5py==2.8.0
httplib2==0.12.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonpickle==1.4.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
labmaze==1.0.3
lockfile==0.12.2
lxml==4.6.1
lz4==3.0.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.1
-e git+git://github.com/aravindr93/mjrl@57c4a8f28caab18c4253c6f47da6d4667b5960a3#egg=mjrl
mock==4.0.2
more-itertools==8.2.0
mujoco-py==1.50.1.68
multipledispatch==0.6.0
# Editable install with no version control (multiworld==0.0.0)
-e /NAS2020/Share/mcchen/mopo/multiworld
networkx==2.2
numpy==1.18.4
opencv-python-headless==4.2.0.34
packaging==20.3
pandas==1.0.5
path==15.0.0
Pillow==7.1.2
plotly==1.9.6
pluggy==0.13.1
protobuf==3.13.0
psutil==5.7.0
py==1.8.1
py-spy==0.3.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycosat==0.6.3
pycparser==2.20
pygame==1.9.6
pyglet==1.5.5
PyOpenGL==3.1.5
pyOpenSSL==18.0.0
pyparsing==2.4.7
PySocks==1.6.8
pytest==5.4.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.0.1
PyYAML==5.3.1
ray==0.6.4
redis==3.5.0
requests==2.23.0
rsa==3.4.2
s3transfer==0.1.13
scikit-image==0.14.2
scikit-learn==0.20.1
scipy==1.4.1
seaborn==0.11.0
# Serializable==0.1.0
# Not sure why
-e git+https://github.com/hartikainen/serializable.git@76516385a3a716ed4a2a9ad877e2d5cbcf18d4e6#egg=Serializable
setproctitle==1.1.10
six==1.14.0
smmap==3.0.4
smmap2==3.0.1
tensorboard==1.13.1
tensorboardX==1.8
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-probability==0.6.0
termcolor==1.1.0
toolz==0.11.1
# torch @ file:///newNAS/Share/mhliu/torch-1.3.0-cp36-cp36m-manylinux1_x86_64.whl
# torchvision @ file:///newNAS/Share/mhliu/torchvision-0.4.1-cp36-cp36m-linux_x86_64.whl
tqdm==4.32.1
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
zipp==3.1.0
